+++ app/javascript/shared/httpClient.js
