+++ app/javascript/shared/i18n-js.js
@@ -1,0 +1,5 @@
+import I18n from 'i18n-js';
+
+window.I18n = I18n;
+
+export default I18n;
