+++ app/javascript/shared/wait.js
