+++ app/views/cars/index.html.erb
@@ -1,1 +1,2 @@
+<% content_for :page_title, 'List' %>
 <%= react_component 'Car' %>
