+++ app/views/layouts/application.html.erb
@@ -2,7 +2,12 @@
 <html>
   <head>
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-    <title><%= t('page_title') %></title>
+      <title>
+        Cars
+        <% if content_for?(:page_title) %>
+          | <%= yield :page_title %>
+        <% end %>
+      </title>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
